improve to run script
